refactor(st-09068): modularize caching middleware
