+++ packages/core/src/langgraph/middleware/caching-entry.ts
@@ -1,0 +1,13 @@
+import type { CacheEntry } from './caching-types.js';
+
+export function isFreshCacheEntry<T>(entry: CacheEntry<T> | undefined, ttl: number): entry is CacheEntry<T> {
+  if (!entry) {
+    return false;
+  }
+
+  return Date.now() - entry.timestamp < ttl;
+}
+
+export function createCachedErrorResult<State>(error: Error): Partial<State> {
+  return { error: error.message } as unknown as Partial<State>;
+}
