+++ planning/features/09-solid-micro-refactors-feature-plan.md
@@ -2,8 +2,8 @@
 
 **Epic Range:** EP-09 through EP-09
 **Status:** In Progress
-**Last Updated:** 2026-06-16
-**Active Story:** ST-09068 - Modularize LangGraph Caching Middleware and Tests (Ready)
+**Last Updated:** 2026-06-17
+**Active Story:** ST-09068 - Modularize LangGraph Caching Middleware and Tests (In Progress)
 
 ---
